Move Token to model-package
This change introduces a new package which will hold multiple heterogenous models
