+++ src/main/java/de/upb/ds/surnia/preprocessing/tasks/TaskInterface.java
@@ -1,6 +1,6 @@
 package de.upb.ds.surnia.preprocessing.tasks;
 
-import de.upb.ds.surnia.preprocessing.Token;
+import de.upb.ds.surnia.preprocessing.model.Token;
 import java.util.List;
 
 public interface TaskInterface {
