+++ src/main/java/de/upb/ds/surnia/queries/QueryParameterReplacer.java
@@ -1,6 +1,6 @@
 package de.upb.ds.surnia.queries;
 
-import de.upb.ds.surnia.preprocessing.Token;
+import de.upb.ds.surnia.preprocessing.model.Token;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
